logger: move logger to cutil/logger
