+++ cmd/delete.go
@@ -17,7 +17,7 @@ package cmd
 import (
 	"fmt"
 	"github.com/kris-nova/kubicorn/cutil"
-	"github.com/kris-nova/kubicorn/logger"
+	"github.com/kris-nova/kubicorn/cutil/logger"
 	"github.com/kris-nova/kubicorn/state"
 	"github.com/kris-nova/kubicorn/state/fs"
 	"github.com/pkg/errors"
