+++ cmd/root.go
@@ -16,7 +16,7 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/kris-nova/kubicorn/logger"
+	"github.com/kris-nova/kubicorn/cutil/logger"
 	"github.com/spf13/cobra"
 	"os"
