+++ cutil/scp/scp.go
@@ -16,7 +16,7 @@ package scp
 
 import (
 	"fmt"
-	"github.com/kris-nova/kubicorn/logger"
+	"github.com/kris-nova/kubicorn/cutil/logger"
 	"github.com/pkg/sftp"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/terminal"
