+++ state/fs/impl.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"github.com/ghodss/yaml"
 	"github.com/kris-nova/kubicorn/apis/cluster"
-	"github.com/kris-nova/kubicorn/logger"
+	"github.com/kris-nova/kubicorn/cutil/logger"
 	"github.com/kris-nova/kubicorn/state"
 	"io"
 	"io/ioutil"
